django_cab_availability/requirements.txt
Django
djangorestframework
pkg-resources
pytz
sqlparse